backend/AWS/aws_requirements.txt
fastapi
uvicorn
boto3
opensearch-py
requests-aws4auth
pydantic
python-multipart
python-dotenv
